fix it! + add TODO
